pyproject.toml
[tool.poetry]
name = "test-02"
version = "0.1.0"
description = ""
authors = ["Husein Naimov <huseinnaimov72@gmail.com>"]
readme = "README.md"
packages = [{include = "test_02"}]

[tool.poetry.dependencies]
python = "^3.11.7"
alembic = "1.13.1"
annotated-types = "0.6.0"
anyio = "4.2.0"
async-timeout = "4.0.3"
asyncpg = "0.29.0"
click = "8.1.7"
colorama = "0.4.6"
fastapi = "0.108.0"
greenlet = "3.0.3"
h11 = "0.14.0"
idna = "3.6"
iniconfig = "2.0.0"
mako = "1.3.2"
markupsafe = "2.1.5"
packaging = "23.2"
pluggy = "1.4.0"
psycopg2-binary = "2.9.9"
pydantic-core = "2.16.2"
pydantic = "2.6.1"
pytest-asyncio = "0.23.4"
pytest = "7.4.4"
sniffio = "1.3.0"
sqlalchemy = "2.0.23"
starlette = "0.32.0.post1"
typing-extensions = "4.9.0"
uvicorn = {version = "0.25.0", extras = ["standart"]}
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
